move incoming to commands.incoming
