+++ bot.py
@@ -36,15 +36,6 @@ dispatcher = updater.dispatcher
 
 
 
-@send_typing_action
-@log
-def incoming(update, context):
-    '''main function that deals with incoming messages that are the meat and potatos of this bot'''
-    if not update.effective_message.text:
-        return
-    responses = main(update, context)
-    for response in responses:
-        send(response, update, context)
 
 
 @send_typing_action
